:recycle: Rename download signed acknowledgement component to signed confidentiality acknowledgement.
Updates component file name, import, and usage for naming consistency.
